requirements_linux.txt
absl-py==2.1.0
aiohttp==3.9.3
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
anyio==4.3.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.2.0
Babel==2.14.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.4
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
comm==0.2.1
command-not-found==0.3
contourpy==1.2.0
cramjam==2.8.2
cryptography==3.4.8
cycler==0.12.1
dbus-python==1.2.18
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
distro==1.7.0
distro-info==1.1+ubuntu0.1
easydict==1.12
einops==0.8.0
exceptiongroup==1.2.1
executing==2.0.1
fastjsonschema==2.19.1
fastparquet==2024.2.0
filelock==3.9.0
fonttools==4.49.0
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2023.4.0
gdown==5.1.0
graphviz==0.20.1
grpcio==1.62.0
h11==0.14.0
httpcore==1.0.4
httplib2==0.20.2
httpx==0.27.0
hydra-core==1.3.2
idna==3.6
igraph==0.11.4
importlib-metadata==4.6.4
ipykernel==6.29.2
ipympl==0.9.4
ipython==8.22.1
ipython-genutils==0.2.0
ipywidgets==8.1.2
isoduration==20.11.0
jedi==0.19.1
jeepney==0.7.1
Jinja2==3.1.2
joblib==1.3.2
json5==0.9.17
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.9.0
jupyter-lsp==2.2.2
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyter_server==2.12.5
jupyter_server_terminals==0.5.2
jupyterlab==4.1.2
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.3
jupyterlab_widgets==3.0.10
keyring==23.5.0
kiwisolver==1.4.5
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lightning==2.2.0.post0
lightning-thunder==0.1.0
lightning-utilities==0.10.1
looseversion==1.3.0
lz4==4.3.3
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.3
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
more-itertools==8.10.0
mpmath==1.3.0
multidict==6.0.5
nbclient==0.9.0
nbconvert==7.16.1
nbformat==5.9.2
nest-asyncio==1.6.0
netifaces==0.11.0
networkx==3.2.1
notebook==7.1.0
notebook_shim==0.2.4
numpy==1.26.3
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.0
omegaconf==2.3.0
onnx==1.16.0
onnxscript==0.1.0.dev20240331
opt-einsum==3.3.0
optree==0.11.0
overrides==7.7.0
packaging==23.2
pandas==2.2.1
pandocfilters==1.5.1
parso==0.8.3
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.2.0
plotly==5.22.0
prometheus_client==0.20.0
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
py4j==0.10.9.7
pyarrow==15.0.0
pycparser==2.21
Pygments==2.17.2
PyGObject==3.42.1
PyJWT==2.3.0
pyparsing==3.1.1
PySocks==1.7.1
python-apt==2.4.0+ubuntu2
python-dateutil==2.8.2
python-dotenv==1.0.1
python-json-logger==2.0.7
pytorch-lightning==2.2.0.post0
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.1.2
qtconsole==5.5.1
QtPy==2.4.1
referencing==0.33.0
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.0
rootutils==1.0.7
rpds-py==0.18.0
scikit-learn==1.4.1.post1
scipy==1.12.0
SecretStorage==3.3.1
Send2Trash==1.8.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
stack-data==0.6.3
sympy==1.12
systemd-python==234
tenacity==8.2.3
tensorboard==2.16.2
tensorboard-data-server==0.7.2
terminado==0.18.0
texttable==1.7.0
threadpoolctl==3.3.0
tinycss2==1.2.1
tomli==2.0.1
torch==2.2.2
torch-tb-profiler==0.4.3
torchaudio==2.2.2
torchinfo==1.8.0
torchmetrics==1.3.1
torchtext==0.17.2
torchvision==0.17.2
tornado==6.4
tqdm==4.66.2
traitlets==5.14.1
triton==2.2.0
types-python-dateutil==2.8.19.20240106
typing_extensions==4.8.0
tzdata==2024.1
ubuntu-advantage-tools==8001
ufw==0.36.1
unattended-upgrades==0.1
uri-template==1.3.0
urllib3==2.2.1
wadllib==1.3.6
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
Werkzeug==3.0.1
widgetsnbextension==4.0.10
yarl==1.9.4
zipp==1.0.0
